Fixed: Opening the zim file with kiwix fails to open(not showing the copy/move dialog).
* The issue occurred because the logic for opening ZIM files was inside the KiwixReaderFragment, and if the user was on any other screen while using the application, this code was not triggered. As a result, the copy/move dialog was not shown. To resolve this, we moved the functionality to the KiwixMainActivity.
